feat: Detect either `uv` or `juv` executable (#1)
The extension now detects whether `uv` or `juv` is available,
prioritizing `uvx juv` for execution. This makes it more likely to work
out-of-the-box while and allowing better control over the `juv` version
run in the extension.

Also introduces a `juv.executable` configuration option to let users
explicitly specify the command or path if they prefer to override the
default behavior.

```json
{
  "juv.executable": "/path/to/juv"
}
```